% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_functions.R
\name{add_rownames}
\alias{add_rownames}
\title{Get the last element of a vector}
\usage{
add_rownames(df, col_name)
}
\arguments{
\item{df}{Dataframe}

\item{col_name}{Column name to use as new rownames}
}
\description{
Get the last element of a vector
}
