% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{adjust_samples}
\alias{adjust_samples}
\title{Adjust sample names}
\usage{
adjust_samples(counts, target)
}
\arguments{
\item{counts}{Expression counts. Samples as columns}

\item{target}{Phenotypic information. Samples as rows}
}
\description{
Adjust samples between sample sheet and files for DESeq2
}
