% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_functions.R
\name{analysis_DEseq.looper}
\alias{analysis_DEseq.looper}
\title{Create a loop to call DESeq2}
\usage{
analysis_DEseq.looper(
  OUTPUT_Data_dir_given,
  dfAnnotation.given,
  list_of_cols.given,
  list_of_cmps.given,
  data_type = "mRNA",
  int_threads = 2,
  sign_value.given = 0.05,
  LFC.given = log2(1.2),
  early_return = FALSE,
  comp_ID = NULL,
  localFit = FALSE,
  forceResults = FALSE,
  min_cutoff_to_plot = 3,
  max_cutoff_to_plot = 50,
  shrinkage.given = "apeglm",
  count_table.given = NULL,
  physeq.obj = NULL,
  tax_levels = NULL,
  gene_annot.given = NULL
)
}
\arguments{
\item{OUTPUT_Data_dir_given}{Absolute path to store results}

\item{dfAnnotation.given}{Samplesheet containing metadata information}

\item{list_of_cols.given}{List of columns of interest to subset from metadata}

\item{data_type}{Either mRNA, miRNA or rRNA_16S. Default: mRNA}

\item{int_threads}{Number of threads to use}

\item{sign_value.given}{Adjusted pvalue cutoff. Default=0.05,}

\item{LFC.given}{Log Fold change cutoff. Default=log2(1.2),}

\item{early_return}{Whether to return exploratory results early or not}

\item{comp_ID}{Tag name to include for each comparison}

\item{localFit}{Use a fitType=local for mean dispersion fit in DESeq2}

\item{forceResults}{Boolean to force re-run analysis if already generated in the folder provided}

\item{min_cutoff_to_plot}{Minimun number of genes significant to continue analysis. Default=3}

\item{max_cutoff_to_plot}{Number of genes significant to plot as candidates analysis. Default=50}

\item{shrinkage.given}{LFC shrinkage estimator provided. Available: apeglm, ashr or normal}

\item{count_table.given}{Dataframe/matrix of counts}

\item{physeq.obj}{Phyloseq object to use}

\item{tax_levels}{Taxonomic levels to use. Default: c("Domain", "Phylum", "Class", "Order", "Family", "Genus", "Species")}

\item{gene_annot.given}{Dataframe containing gene annotation (Default: NULL)}

\item{formula_given}{Design formula to use}

\item{coef_n}{Number of the coefficient of results to test (if desired)}

\item{cutoff.given}{add an option to include cutoff when removing Zeros}
}
\description{
Create a loop to call DESeq2
}
