% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_functions.R
\name{analysis_DESeq}
\alias{analysis_DESeq}
\title{DESEQ2 analysis pipeline}
\usage{
analysis_DESeq(
  OUTPUT_Data_dir_given,
  count_table,
  sample_sheet_given,
  list_of_cols,
  formula_given,
  int_threads = 2,
  sign_value.given = 0.05,
  LFC.given = log2(1.2),
  coef_n = NA,
  early_return = FALSE,
  comp_ID = NULL,
  cutoff.given = 0.9,
  localFit = FALSE,
  forceResults = FALSE,
  gene.annot = NULL,
  shrinkage.given = "apeglm"
)
}
\arguments{
\item{OUTPUT_Data_dir_given}{Absolute path to store results}

\item{count_table}{Dataframe/matrix of counts}

\item{sample_sheet_given}{Samplesheet containing metadata information}

\item{formula_given}{Design formula to use}

\item{int_threads}{Number of threads to use}

\item{sign_value.given}{Adjusted pvalue cutoff. Default=0.05,}

\item{LFC.given}{Log Fold change cutoff. Default=log2(1.2),}

\item{coef_n}{Number of the coefficient of results to test \link{if desired}}

\item{early_return}{Whether to return exploratory results early or not}

\item{comp_ID}{Tag name to include for each comparison}

\item{cutoff.given}{add an option to include cutoff when removing Zeros}

\item{localFit}{Use a fitType=local for mean dispersion fit in DESeq2}

\item{forceResults}{Boolean to force re-run analysis if already generated in the folder provided}

\item{gene.annot}{Dataframe containing gene annotation (Default: NULL)}

\item{shrinkage.given}{LFC shrinkage estimator provided. Available: apeglm, ashr or normal}
}
\description{
When running DESeq2 you usually add multiple terms to the matrix design. Test the effect of them
}
