% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_functions.R
\name{analysis_DESeq_phyloseq}
\alias{analysis_DESeq_phyloseq}
\title{Call DESeq2 from phyloseq object}
\usage{
analysis_DESeq_phyloseq(
  OUTPUT_Data_dir_given,
  physeq.obj,
  sample_sheet_given,
  list_of_cols,
  formula_given,
  int_threads = 2,
  sign_value.given = 0.05,
  LFC.given = log2(1.2),
  coef_n = NA,
  early_return = FALSE,
  comp_ID = NULL,
  cutoff.given = 0.9,
  localFit = FALSE,
  forceResults = FALSE,
  min_cutoff_to_plot = 3,
  max_cutoff_to_plot = 50,
  gene.annot = NULL,
  data_type = "rRNA_16S",
  shrinkage.given = "apeglm",
  tax_levels = c("Domain", "Phylum", "Class", "Order", "Family", "Genus", "Species")
)
}
\arguments{
\item{OUTPUT_Data_dir_given}{Absolute path to store results}

\item{physeq.obj}{Phyloseq object to use}

\item{sample_sheet_given}{Samplesheet containing metadata information}

\item{list_of_cols}{List of columns of interest to subset from metadata}

\item{formula_given}{Design formula to use}

\item{int_threads}{Number of threads to use}

\item{sign_value.given}{Adjusted pvalue cutoff. Default=0.05,}

\item{LFC.given}{Log Fold change cutoff. Default=log2(1.2),}

\item{coef_n}{Number of the coefficient of results to test (if desired)}

\item{early_return}{Whether to return exploratory results early or not}

\item{comp_ID}{Tag name to include for each comparison}

\item{cutoff.given}{add an option to include cutoff when removing Zeros}

\item{localFit}{Use a fitType=local for mean dispersion fit in DESeq2}

\item{forceResults}{Boolean to force re-run analysis if already generated in the folder provided}

\item{min_cutoff_to_plot}{Minimun number of genes significant to continue analysis. Default=3}

\item{max_cutoff_to_plot}{Number of genes significant to plot as candidates analysis. Default=50}

\item{gene.annot}{Dataframe containing gene annotation (Default: NULL)}

\item{data_type}{Either mRNA, miRNA or rRNA_16S. Default: rRNA_16S}

\item{shrinkage.given}{LFC shrinkage estimator provided. Available: apeglm, ashr or normal}

\item{tax_levels}{Taxonomic levels to use. Default: c("Domain", "Phylum", "Class", "Order", "Family", "Genus", "Species")}
}
\description{
Call DESeq2 from phyloseq object
}
