% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAcyto_functions.R
\name{ASCAT_LOH}
\alias{ASCAT_LOH}
\title{Compute LOH from count of alleles}
\usage{
ASCAT_LOH(nMajor, nMinor)
}
\arguments{
\item{nMajor}{Number copies of major allele}

\item{nMinor}{Number copies of minor allele}
}
\description{
Compute LOH from count of alleles
}
