% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_functions.R
\name{call_dmrcate}
\alias{call_dmrcate}
\title{Call DMRcate and previous steps}
\usage{
call_dmrcate(
  m_values.noSNPs.given,
  my_cmp,
  my_dir,
  metadata.given,
  design.cmp,
  fdr.given = 0.05,
  coef.given = 2,
  lambda.given = 1000,
  C.given = 2,
  paired_variable = NULL
)
}
\arguments{
\item{m_values.noSNPs.given}{}

\item{my_cmp}{}

\item{my_dir}{}

\item{metadata.given}{}

\item{design.cmp}{}

\item{fdr.given}{}

\item{coef.given}{}

\item{lambda.given}{}

\item{C.given}{}

\item{paired_variable}{}
}
\description{
Call DMRcate and previous steps
}
