% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_functions.R
\name{call_dmrcate_comparisons}
\alias{call_dmrcate_comparisons}
\title{Call DMRcate for several comparisons}
\usage{
call_dmrcate_comparisons(
  sample_metadata.given,
  var_name,
  dir_folder.given,
  m_values.noSNPs.given,
  lambda.given = 1000,
  C.given = 2,
  fdr.given = 0.05,
  paired_variable = NULL,
  other_var_name = NULL
)
}
\arguments{
\item{sample_metadata.given}{Dataframe with information}

\item{var_name}{Name for the comparison variable}

\item{dir_folder.given}{Folder to store results}

\item{m_values.noSNPs.given}{Dataframe with probes as rownames, m values with no SNPs (as a result of DMRcate::rmSNPandCH)}

\item{lambda.given}{dmrcate parameter. Gaussian kernel bandwidth for smoothed-function estimation. Default: 1000}

\item{C.given}{dmrcate parameter. Scaling factor for bandwidth. Gaussian kernel is calculated where lambda/C = sigma. Default: 2}

\item{fdr.given}{False discovery rate cutoff. Default 0.25.}
}
\description{
Call DMRcate for several comparisons
}
