% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_functions.R
\name{call_limma_comparisons}
\alias{call_limma_comparisons}
\title{Call limma for several comparisons}
\usage{
call_limma_comparisons(
  sample_metadata.given,
  var_name,
  dir_folder.given,
  values_df,
  fdr.given = 0.05,
  EPICannotation.given,
  bvalues,
  paired_variable = NULL,
  other_var_name = NULL
)
}
\arguments{
\item{sample_metadata.given}{Dataframe with information}

\item{var_name}{Name for the comparison variable}

\item{dir_folder.given}{Folder to store results}

\item{values_df}{Dataframe with probes as rownames, m values with no SNPs (as a result of DMRcate::rmSNPandCH)}

\item{fdr.given}{False discovery rate cutoff. Default 0.05.}

\item{EPICannotation.given}{EPIC annotation dataframe}

\item{bvalues}{Beta values dataframe}

\item{paired_variable}{If pair samples, include this variable}

\item{other_var_name}{Other variable to include in the design}
}
\description{
Call limma for several comparisons
}
