% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPICv2_functions.R
\name{check.background}
\alias{check.background}
\title{Background intensity plotter}
\usage{
check.background(
  qcs.df.given,
  mean_oob_grn = "mean_oob_grn",
  mean_oob_red = "mean_oob_red",
  Sample_Name = "Sample_Name",
  xlim_vec = c(100, 1000),
  ylim_vec = c(100, 1000),
  color_cat
)
}
\arguments{
\item{qcs.df.given}{Dataframe from getDataF_sesameQC_calcStats}

\item{mean_oob_grn}{Column with mean oob green intensity. Default: mean_oob_grn}

\item{mean_oob_red}{Column with mean oob red intensity. Default: mean_oob_red}

\item{Sample_Name}{Column with sample ids}

\item{xlim_vec}{Axis limits. Default: 100-1000}

\item{ylim_vec}{Axis limits. Default: 100-1000}

\item{color_cat}{Column to use for coloring samples}
}
\value{
ggplot2 object
}
\description{
Create Out-of-bag (OOB) red/green ratio plot from data frame containing values.
Data frame can be generated from getDataF_sesameQC_calcStats or any other source,
use appropriate name for columns conveniently
}
