% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPICv2_functions.R
\name{check.mean_intensity}
\alias{check.mean_intensity}
\title{Intensity plotter}
\usage{
check.mean_intensity(
  qcs.df.given,
  intensity,
  y_label,
  fill_cat,
  Sample_Name = "Sample_Name"
)
}
\arguments{
\item{qcs.df.given}{Dataframe from getDataF_sesameQC_calcStats}

\item{intensity}{Intensity column name in dataframe to plot for each sample}

\item{y_label}{Axis Y label sentence}

\item{fill_cat}{Column to use for coloring samples}

\item{Sample_Name}{Column with sample ids}
}
\value{
ggplot2 object
}
\description{
Intensity plotter
}
