% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPICv2_functions.R
\name{check.na_probes}
\alias{check.na_probes}
\title{Missing values NA plotter}
\usage{
check.na_probes(
  qcs.df.given,
  na2check,
  y_label,
  fill_cat,
  Sample_Name = "Sample_Name"
)
}
\arguments{
\item{qcs.df.given}{Dataframe from getDataF_sesameQC_calcStats}

\item{na2check}{Missing values columns column name in dataframe to plot for each sample:}

\item{y_label}{Axis Y label sentence}

\item{fill_cat}{Column to use for coloring samples}

\item{Sample_Name}{Column with sample ids}
}
\value{
ggplot2 object
}
\description{
Missing values NA plotter
}
