% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBeads_functions.R
\name{check_covariates}
\alias{check_covariates}
\title{Check for covariates using RnBead object}
\usage{
check_covariates(var2check, rnb.set_given, plot_dir)
}
\arguments{
\item{var2check}{Suspected variable to test for Covariate}

\item{rnb.set_given}{RnBeads rnb.set object}

\item{plot_dir}{Folder to save plot results}
}
\description{
This functions checks for covariates at cpg and promoters
}
