% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{check_rank_design}
\alias{check_rank_design}
\title{Check the rank of a design matrix}
\usage{
check_rank_design(formula2test, data.df)
}
\arguments{
\item{formula2test}{String with formula to check}

\item{data.df}{Sample sheet dataframe}
}
\description{
When running DESeq2 you need a design matrix, check the rank of it first
}
