% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{check_reduced_LRT}
\alias{check_reduced_LRT}
\title{Check the effect of reducing one term from a DESEQ2 design}
\usage{
check_reduced_LRT(
  dds_obj.given,
  formula_given,
  comp.folder.given,
  compID.given,
  dfAnnotation.given,
  int_threads = 2
)
}
\arguments{
\item{dds_obj.given}{DESeq2 object (DESeqDataSet)}

\item{formula_given}{Formula to substract for formula in dds_obj.given}

\item{comp.folder.given}{Absolute path to store results}

\item{compID.given}{Tag name to include for each comparison}

\item{dfAnnotation.given}{Dataframe with useful metadata to include}
}
\description{
Check reducing effect using LRT method.
}
