% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{check_terms_matrix}
\alias{check_terms_matrix}
\title{Check the effect of variables in matrix design}
\usage{
check_terms_matrix(
  sampleSheet.given,
  countsGiven,
  list.terms,
  red.formula.given,
  compID.given,
  dfAnnotation.given,
  comp.folder.given,
  int_threads = 2
)
}
\arguments{
\item{sampleSheet.given}{Samplesheet containing metadata information}

\item{countsGiven}{Dataframe/matrix of counts}

\item{list.terms}{List of terms from samplesheet to include in design matrix}

\item{red.formula.given}{Design formula to use as naive and in reduction}

\item{compID.given}{Tag name to include for each comparison}

\item{dfAnnotation.given}{Dataframe with useful metadata to include}

\item{comp.folder.given}{Absolute path to store results}

\item{int_threads}{Number of threads to use}
}
\description{
When running DESeq2 you usually add multiple terms to the matrix design. Test the effect of them
}
