% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_functions.R
\name{colnames_GRanges}
\alias{colnames_GRanges}
\title{Get column names for a GRanges object}
\usage{
colnames_GRanges(granges_given)
}
\arguments{
\item{granges_given}{}
}
\description{
Get column names for a GRanges object
}
