% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convertDate}
\alias{convertDate}
\title{Convert Spanish date: 14/5/2022 into English american format: 2022-5-14}
\usage{
convertDate(date_str)
}
\arguments{
\item{date_str}{String date to convert}
}
\description{
Convert Spanish date: 14/5/2022 into English american format: 2022-5-14
}
