% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_col_palette}
\alias{create_col_palette}
\title{Assigns color codes for samples}
\usage{
create_col_palette(columnGiven, levels_given, palette_given = "Paired")
}
\arguments{
\item{columnGiven}{Column dataframe given to use as class for each color}

\item{levels_given}{Orderered or unordered levels to use}

\item{palette_given}{Name of the palette to use: Paired as default.}
}
\description{
Create palette color codes for each sample given using RColorBrewer given a variable
}
