% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{create_DE_plots}
\alias{create_DE_plots}
\title{Create DE gene plots for several metadata variables}
\usage{
create_DE_plots(
  outdir_path,
  metadata_given,
  list_of_cols,
  sign.data,
  comp_name,
  numerator,
  denominator,
  gene.annot.df = NULL,
  max_cutoff_to_plot = 50,
  data_type = "mRNA",
  n_cores = 2,
  input_type = "DESeq2"
)
}
\arguments{
\item{outdir_path}{Output dirname}

\item{metadata_given}{Samplesheet dataframe}

\item{list_of_cols}{Columns to use for plotting genes}

\item{sign.data}{Signficant dataframe to use}

\item{comp_name}{Comparison of interest}

\item{numerator}{Group 1, used as numerator}

\item{denominator}{Group 2, used as denominator or reference in the comparison}

\item{gene.annot.df}{Dataframe with annotation, either for miRNA or mRNA containing parent and variant or hgnc_symbol, respectively. Default: NULL}

\item{max_cutoff_to_plot}{Maximun number of genes to plot. Default: 50}

\item{data_type}{Either mRNA, miRNA or rRNA_16S. Default: mRNA}

\item{input_type}{Choose either DESeq2 or limma to identify the format of the results.}

\item{outdir_name}{Name of the folder to create. Default: boxplot_DE}
}
\description{
Create DE gene plots for several metadata variables
}
