% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAcyto_functions.R
\name{create_matrix_overlaps}
\alias{create_matrix_overlaps}
\title{Create matrix of region overlaps}
\usage{
create_matrix_overlaps(GR.list_given)
}
\arguments{
\item{GR.list_given}{List of Genomic ranges objects per sample}
}
\value{
matrix
}
\description{
This functions takes a genomicRanges object and using regioneR::numOverlaps creates a matrix of
overlapping regions
}
