% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAcyto_functions.R
\name{create_matrix_overlaps_by_seqnames}
\alias{create_matrix_overlaps_by_seqnames}
\title{Create matrix of region overlaps (by sequence)}
\usage{
create_matrix_overlaps_by_seqnames(GR.list_given, list_of_chrs = c(1:22))
}
\arguments{
\item{GR.list_given}{List of Genomic ranges objects per sample}

\item{list_of_chrs}{List of seq IDs to use, by default chromosomes 1-22.}
}
\description{
This functions takes a genomicRanges object and using regioneR::numOverlaps creates a matrix of
overlapping regions by sequence.
}
