% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_functions.R
\name{create_mult_perm_test}
\alias{create_mult_perm_test}
\title{Create multiple permutation test}
\usage{
create_mult_perm_test(
  sample_name,
  GSEA.genes.GR_given,
  list_GRs_list,
  ntimes = 1000,
  verbose = TRUE
)
}
\arguments{
\item{sample_name}{Name of sample to use}

\item{GSEA.genes.GR_given}{GRs of genes to intersect}

\item{list_GRs_list}{List of GRs to use}

\item{ntimes}{Number of times for regioneR}

\item{verbose}{T/F for verbosity}
}
\description{
Create multiple permutation test
}
