% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_functions.R
\name{create_overlaps}
\alias{create_overlaps}
\title{Create overlaps for a GR vs. list of GRs}
\usage{
create_overlaps(GR_given, GR.list_given, mc.cores = 2)
}
\arguments{
\item{GR_given}{}

\item{GR.list_given}{}

\item{mc.cores}{}
}
\description{
Create overlaps for a GR vs. list of GRs
}
