% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_pcas}
\alias{create_pcas}
\title{Create PCA analysis and plots}
\usage{
create_pcas(
  values.df,
  tag_to_use,
  plots_dir_given,
  sample_sheet.given,
  col_of_interest,
  subset_n = NULL
)
}
\arguments{
\item{values.df}{Numerical values to use for the PCA analysis}

\item{tag_to_use}{Name to include in the file to save results}

\item{plots_dir_given}{Path to store PDF with plots}

\item{sample_sheet.given}{Metadata for each sample (as rownames) to colour PCA results}

\item{col_of_interest}{List of columns from sample sheet to use}

\item{subset_n}{Number of rows to use from original values.df. Default is to use all}
}
\description{
Given a dataframe create PCA and then produce multiple plots using PC1 & PC2 and
a list of columns to colour samples
}
