% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPICv2_functions.R
\name{create_rank_plots}
\alias{create_rank_plots}
\title{Ranking plots per sample}
\usage{
create_rank_plots(df.given, set2plot, nameSet, pdf_file.dir)
}
\arguments{
\item{df.given}{Dataframe from getDataF_sesameQC_calcStats}

\item{set2plot}{Name of columns to use. See get_keys.qcs() for options}

\item{nameSet}{Name to use for saving PDF plots}

\item{pdf_file.dir}{Path to store files}
}
\description{
Ranking plots per sample
}
