% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpSetR_functions.R
\name{create_upset}
\alias{create_upset}
\title{Creates UpSetR plot}
\usage{
create_upset(data_dir, pattern2search)
}
\arguments{
\item{data_dir}{Absolute path to folder containing files with e.g. comp1-comp2_down.whatever.txt}

\item{pattern2search}{Pattern to include in regex to search file names. e.g. down, complex.isomir, etc}
}
\value{
List of plot and data containing data for each set
}
\description{
Get information of interest to produce UpSetR plots from a folder containing files and return plot generated
}
