% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPICv2_functions.R
\name{db_exploratory}
\alias{db_exploratory}
\title{Dye bias exploratory}
\usage{
db_exploratory(SigDF_given, sample_name, sample_dir, tag_name = "raw_")
}
\arguments{
\item{SigDF_given}{SigDF generated}

\item{sample_name}{Sample ID}

\item{sample_dir}{Folder to store results}

\item{tag_name}{Tag to add to the files generated}
}
\description{
Sesame Dye bias correction by matching green and red to mid point.
This function compares the Type-I Red probes and Type-I Grn probes and generates
and mapping to correct signal of the two channels to the middle.
}
