% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delete.na}
\alias{delete.na}
\title{Remove NAs in dataframe}
\usage{
\method{delete}{na}(DF, n = 0)
}
\arguments{
\item{DF}{Dataframe provide}

\item{n}{Number of NAs allowed for each row.}
}
\description{
Filters out dataframe according the amount of NAs allowed
}
