% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_functions.R
\name{delete_na}
\alias{delete_na}
\title{Remove NAs in dataframe}
\usage{
delete_na(DF, n = 0)
}
\arguments{
\item{DF}{Dataframe provide}

\item{n}{Number of NAs allowed for each row.}
}
\description{
Filters out dataframe according the amount of NAs allowed
}
