% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_functions.R
\name{DESeq2_HCGB_function}
\alias{DESeq2_HCGB_function}
\title{DESEq2 main function}
\usage{
DESeq2_HCGB_function(
  dds_object,
  coef_n,
  comp_name,
  comp_ID = "comp1",
  numerator = "example1",
  denominator = "example2",
  OUTPUT_Data_dir,
  df_treatment_Ind,
  list_of_cols,
  threads = 2,
  sign_value.given = 0.05,
  LFC.given = log2(1.2),
  min_cutoff_to_plot = 3,
  max_cutoff_to_plot = 50,
  gene.annot.df = NULL,
  data_type = "mRNA",
  forceResults = FALSE,
  shrinkage = "apeglm"
)
}
\arguments{
\item{dds_object}{DESeq2 object}

\item{coef_n}{Coefficient number obtain using resultsName(dds)}

\item{comp_name}{Name of the comparison}

\item{comp_ID}{ID tag of the comparison}

\item{numerator}{Name of the numerator comparison}

\item{denominator}{Name of the denominator comparison}

\item{OUTPUT_Data_dir}{Folder path to store results}

\item{df_treatment_Ind}{Dataframe containing additional information for each sample}

\item{list_of_cols}{Set of columns with important information in df_treatment_ind}

\item{threads}{Number of CPUs to use (Default: 2).}

\item{sign_value.given}{Adjusted pvlaue cutoff. Default=0.05,}

\item{LFC.given}{Log Fold change cutoff. Default=log2(1.2),}

\item{min_cutoff_to_plot}{Minimun number of genes significant to continue analysis. Default=3}

\item{max_cutoff_to_plot}{Number of genes significant to plot as candidates analysis. Default=50}

\item{gene.annot.df}{Dataframe containing gene annotation (Default: NULL)}

\item{forceResults}{Boolean to force re-run analysis if already generated in the folder provided}
}
\description{
This functions creates the main analysis given a DDS and comparison to test
}
