% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_functions.R
\name{dim_GRanges}
\alias{dim_GRanges}
\title{Get dimensions for a GRanges object}
\usage{
dim_GRanges(granges_given)
}
\arguments{
\item{granges_given}{}
}
\description{
Get dimensions for a GRanges object
}
