% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{discard_0_counts}
\alias{discard_0_counts}
\title{Discard empty counts}
\usage{
discard_0_counts(countsF, cutoff = 0.9)
}
\arguments{
\item{countsF}{Count file for DESeq}

\item{cutoff}{Cutoff for discarding rows}
}
\description{
This functions prepares countsfile for DESeq discarding rows with
many 0 for all samples. It uses the given cutoff, to discard rows containing
greater than a cutoff in percentage of 0's. Default cutoff value is 0.9
}
