% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{discard_lowCounts}
\alias{discard_lowCounts}
\title{Discard low count in a dds object}
\usage{
discard_lowCounts(dds_object, min_count = 10)
}
\arguments{
\item{dds_object}{DESeq2 object}

\item{min_count}{integer for minimum count cutoff \link{Default=10}.}
}
\description{
This functions discard low counts in a DESeq object.
Uses 10 as default.
}
