% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{discard_lowCounts_df}
\alias{discard_lowCounts_df}
\title{Discard low counts in dataframe}
\usage{
discard_lowCounts_df(df_given, min_count = 10)
}
\arguments{
\item{df_given}{Dataframe to pase}

\item{min_count}{Minimum count to use as cutoff \link{Default: 10}.}
}
\description{
This functions discard low counts in a dataframe with any non-numeric column
}
