% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_functions.R
\name{distance_ord_plot}
\alias{distance_ord_plot}
\title{Ordination plots}
\usage{
distance_ord_plot(
  dist_method,
  phyloseq_obj,
  color_given,
  shape_given,
  ordination_method = "MDS"
)
}
\arguments{
\item{dist_method}{phyloseq::distance method to use (Available in physeq::distanceMethodList())}

\item{phyloseq_obj}{Phyloseq object to plot}

\item{color_given}{Metadata column to use for color}

\item{shape_given}{Metadata column to use for shape}

\item{ordination_method}{Ordination method. Options: c("DCA", "CCA", "RDA", "CAP", "DPCoA", "NMDS", "MDS", "PCoA"). Default: MDS}
}
\description{
Ordination plots
}
