% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen2phyloseq.R
\name{dragen2phyloseq}
\alias{dragen2phyloseq}
\title{dragen2phyloseq: import illumina DRAGEN metagenomics into phyloseq object}
\usage{
dragen2phyloseq(sample_sheet, col_name = "id", col_filepath = "abs_path")
}
\arguments{
\item{sample_sheet}{Metadata information stored as a dataframe for later phyloseq. This dataframe must contain a column with sample ids and another with each file sample absolute path}

\item{col_name}{Column name to use for sample names. Default: id}

\item{col_filepath}{Column name to use for absolute file path. Default: abs_path}
}
\description{
dragen2phyloseq: import illumina DRAGEN metagenomics into phyloseq object
}
