% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBeads_functions.R
\name{dreduction_function}
\alias{dreduction_function}
\title{Check for covariates using RnBead object}
\usage{
dreduction_function(
  rnb.set_given,
  vars_dreduction,
  methods_dreduction,
  pdf_file
)
}
\arguments{
\item{rnb.set_given}{RnBeads rnb.set object}

\item{var2check}{Suspected variable to test for Covariate}

\item{plot_dir}{Folder to save plot results}
}
\description{
This functions checks for covariates at cpg and promoters
}
