% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_functions.R
\name{enricher_loop}
\alias{enricher_loop}
\title{Create gene ontology enrichment for a given dataset}
\usage{
enricher_loop(
  list_of_sets,
  geneUniverse,
  GO_folder,
  tag2use,
  org.Db2use = "org.Hs.eg.db",
  ont_types = c("CC", "BP", "MF"),
  keyType.given = "SYMBOL"
)
}
\arguments{
\item{list_of_sets}{List of dataset to use}

\item{geneUniverse}{List of total genes to use as "Universe"}

\item{GO_folder}{Folder to store results}

\item{tag2use}{Tag name to include}

\item{org.Db2use}{Database to use: Default: 'org.Hs.eg.db'. It should be loaded first: library('org.Hs.eg.db')}

\item{ont_types}{List of ontologies classes to use. Default: c('CC', 'BP', 'MF')}

\item{keyType.given}{SYMBOL as default. See keytypes(org.Hs.eg.db) as an example for putative keys to use}
}
\description{
Gets coordinates for the set of genes desired
}
