% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FGSEA_GSEA}
\alias{FGSEA_GSEA}
\title{FGSEA enrichment analysis}
\usage{
FGSEA_GSEA(
  gene_list_provided,
  myGeneSet,
  title_given = "example",
  nproc_given = 2
)
}
\arguments{
\item{gene_list_provided}{Named list ranked by either pvalue, logFC, padj}

\item{myGeneSet}{Gene set data as a list of lists}

\item{title_given}{Title to include in the ggplot generated}

\item{nproc_given}{Number of threads to use}
}
\description{
Creates a gene set enrichment analysis using fgsea a given table by the given column name. It requires a GENE_SYMBOL column to include gene IDs.
}
