% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_functions.R
\name{FGSEA_GSEA_loop}
\alias{FGSEA_GSEA_loop}
\title{Create GSEA loop}
\usage{
FGSEA_GSEA_loop(
  table_annot,
  folder_out,
  name_given,
  nproc_given,
  dataSet.list,
  GENE_SYMBOL.col = "GENE_SYMBOL",
  ranker.list = c("log2FoldChange", "padj"),
  padj.thres = 0.25
)
}
\arguments{
\item{table_annot}{table to get results}

\item{folder_out}{folder to store results}

\item{name_given}{name for the comparison}

\item{nproc_given}{threads to use}

\item{dataSet.list}{list of GSEA datasets}

\item{GENE_SYMBOL.col}{name of the hgnc column}

\item{ranker.list}{columns to use for ranking}

\item{padj.thres}{Pvalue adjusted threshold}
}
\description{
Plots GSEA results using ggplot and enrichment score provided
}
