% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_functions.R
\name{filter_bump_table}
\alias{filter_bump_table}
\title{Filter bumps and retain more significant}
\usage{
filter_bump_table(bump_table, clusterL_given = 3, L_given = 3)
}
\arguments{
\item{bump_table}{Bumphunter dataframe}

\item{clusterL_given}{Number of probes to be within each cluster}

\item{L_given}{Numberof probes with the same trending line within each cluster}
}
\description{
This functions filters a bump_table provided. It uses only the cluster L given. It could also use fwer as a method to discard false positives
}
