% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_functions.R
\name{filter_dmpFinder_table}
\alias{filter_dmpFinder_table}
\title{Filter dmpFinder methylation stats}
\usage{
filter_dmpFinder_table(
  dmpData,
  beta_val,
  list1,
  list2,
  pval = 0.05,
  qval = 0.05,
  diff.mean = 0.2
)
}
\arguments{
\item{dmpData}{Dataframe with probes as rownames}

\item{beta_val}{Dataframe containing beta_values for each probe and sample (columns)}

\item{list1}{Vector of samples to test fro group2}

\item{pval}{Pvalue filtering cutoff. Default: 0.05}

\item{qval}{Pvalue adjusted filtering cutoff. Default: 0.05}

\item{diff.mean}{Mean difference (absolute values) for filtering}
}
\description{
This functions obtains a mean beta value for each group of samples and filters based on statistics
}
