% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{filter_signficant_DESEQ}
\alias{filter_signficant_DESEQ}
\title{Filter significant hits from DESEQ2 analysis}
\usage{
filter_signficant_DESEQ(dataF, sign_value = 0.05, LFC = 0.26)
}
\arguments{
\item{dataF}{Dataframe with either normalized values and DESEQ2 values or only DESEQ2.}

\item{sign_value}{Pvalue adjusted cutoff: Default=0.05}

\item{LFC}{Log Fold Change cutoff: Default: 0.26}
}
\description{
When running DESeq2 you usually required significant hits.
}
