% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_functions.R
\name{filter_signficant_limma}
\alias{filter_signficant_limma}
\title{Filter limma DE analysis}
\usage{
filter_signficant_limma(dataF, sign_value = 0.05, LFC = 0.26)
}
\arguments{
\item{sign_value}{Pvalue adjusted cutoff: Default=0.05}

\item{LFC}{Log Fold Change cutoff: Default: 0.26}

\item{normData}{results limma dataframe to use as normalized values}
}
\description{
This functions filters limma DE results
}
