% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{filter_significant_hits}
\alias{filter_significant_hits}
\title{Filter significant hits from differential expression analysis}
\usage{
filter_significant_hits(
  dataF,
  sign_value = 0.05,
  LFC = 0.26,
  input_type = "DESeq2"
)
}
\arguments{
\item{dataF}{Dataframe with either normalized values and DESEQ2 values or only DESEQ2.}

\item{sign_value}{Pvalue adjusted cutoff: Default=0.05}

\item{LFC}{Log Fold Change cutoff: Default: 0.26}

\item{input_type}{Choose either DESeq2 or limma to identify the format of the results.}
}
\description{
When running DESeq2 you usually required significant hits.
}
