% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_functions.R
\name{functional_analysis}
\alias{functional_analysis}
\title{Create functional analysis}
\usage{
functional_analysis(
  alldata,
  dir_given,
  tag_name,
  GSEA.datasets,
  sign.genes,
  nproc_given = 12,
  GENE_SYMBOL.col = "hgnc_symbol",
  ranker.list = c("log2FoldChange", "signed.logPval")
)
}
\arguments{
\item{alldata}{DE dataframe}

\item{dir_given}{Output dir to store results}

\item{tag_name}{Name to use}

\item{GSEA.datasets}{List of datasets from GSEA msigdb}

\item{sign.genes}{List of significant genes}

\item{nproc_given}{Number of threads to use}

\item{GENE_SYMBOL.col}{Gene column name to use: Default: hgnc_symbol}

\item{ranker.list}{For gsea, rank by column Default: c('log2FoldChange', 'signed.logPval')}
}
\description{
Create functional analysis
}
