% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_functions.R
\name{get_abundance.long}
\alias{get_abundance.long}
\title{Get abundances in a long format and save file in folder}
\usage{
get_abundance.long(physeq.obj, results.fold, tag_name = "data")
}
\arguments{
\item{physeq.obj}{Phyloseq object to use}

\item{results.fold}{Results folder to store data in csv and xlsx format}

\item{tag_name}{Tag to add to the name of the files generated}
}
\description{
Get abundances in a long format and save file in folder
}
