% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_functions.R
\name{get_abundant_taxa}
\alias{get_abundant_taxa}
\title{Get most abundant taxa}
\usage{
get_abundant_taxa(physeq_obj, n = 10)
}
\arguments{
\item{physeq_obj}{Phyloseq object to subset}

\item{n}{Number of top taxa to retrieve}
}
\description{
Get most abundant taxa
}
