% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_functions.R
\name{get_all_data_DESeq2}
\alias{get_all_data_DESeq2}
\title{Get results from DESeq2 and normalized data}
\usage{
get_all_data_DESeq2(dds_obj, coef_n, type = "DESeq2")
}
\arguments{
\item{dds_obj}{DESeq2 object (DESeqDataSet)}

\item{coef_n}{Name of the coefficient obtain from DESeq2::resultsNames(dds_obj))}

\item{type}{By default RNA is expected and genes, either EntrezID or ENSEMBL ID is used as ID. You can specify XICRA if miRNA is provided and Gene is the combination of miRNA, variant and isomir (e.g hsa-let-7a-2-3p&iso_add3p:1&iso-23-NLJ18XQZD2). If type XICRA provided, columns is splitted into three new columns (parent, variant and UID)}
}
\description{
When running DESeq2 you usually require to get all statiscal results and normalized data
}
