% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_functions.R
\name{get_beta_mean_vals}
\alias{get_beta_mean_vals}
\title{Get mean beta values for two list of samples}
\usage{
get_beta_mean_vals(table_dat, beta_val, list1, list2)
}
\arguments{
\item{table_dat}{Dataframe with probes as rownames}

\item{beta_val}{Dataframe containing beta_values for each probe and sample (columns)}

\item{list1}{Vector of samples to test fro group2}
}
\description{
This functions obtains a mean beta value for each group of samples
}
