% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_functions.R
\name{get_chr_len}
\alias{get_chr_len}
\title{Return dataframe with chromosome length for hg19 genome}
\usage{
get_chr_len(genome_version = "hg38")
}
\arguments{
\item{genome_version}{Set hg38 or hg19. Default: hg38}
}
\description{
This functions returns chromosome length for hg19 genome
}
