% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{get_comparison_resultsNames}
\alias{get_comparison_resultsNames}
\title{Get names provided in comparison}
\usage{
get_comparison_resultsNames(str_given)
}
\arguments{
\item{str_given}{A string with the comparison. E.g. Sex_male_vs_female}
}
\description{
When running DESeq2 you usually get names from resultsNames() such as var_comp1_vs_comp2 e.g. Sex_male_vs_female. This functions returns the name of the variables and the comparison studied.
}
