% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_functions.R
\name{get_coverage_GR}
\alias{get_coverage_GR}
\title{Get coverage GR or list of GRs

Get coverage for all samples vs. all positions}
\usage{
get_coverage_GR(GR_given)
}
\arguments{
\item{GR_given}{}
}
\description{
Get coverage GR or list of GRs

Get coverage for all samples vs. all positions
}
