% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpSetR_functions.R
\name{get_data_upset}
\alias{get_data_upset}
\title{Get files for UpSetR plot

Get information of interest to produce UpSetR plots from a folder containing files}
\usage{
get_data_upset(data_dir, pattern2search)
}
\arguments{
\item{data_dir}{Absolute path to folder containing files with e.g. comp1-comp2_down.whatever.txt}

\item{pattern2search}{Pattern to include in regex to search file names. e.g. down, complex.isomir, etc}
}
\value{
List of lists containing data for each set
}
\description{
Get files for UpSetR plot

Get information of interest to produce UpSetR plots from a folder containing files
}
