% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_functions.R
\name{get_dim.filt}
\alias{get_dim.filt}
\title{Get numbers rows fulfilling cutoff of filt_length}
\usage{
get_dim.filt(i, filt_length = 10000)
}
\arguments{
\item{i}{}

\item{filt_length}{}
}
\description{
Get numbers rows fulfilling cutoff of filt_length
}
