% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_functions.R
\name{get_diversity_stats}
\alias{get_diversity_stats}
\title{Get richness & diversity statistics}
\usage{
get_diversity_stats(physeq.obj, results.fold)
}
\arguments{
\item{physeq.obj}{Phyloseq object to use}

\item{results.fold}{Results folder to store data in csv and xlsx format}
}
\value{
dataframe
}
\description{
Get richness & diversity statistics
}
