% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen2phyloseq.R
\name{get_dragenData}
\alias{get_dragenData}
\title{Get dragen metagenomics data for all samples}
\usage{
get_dragenData(sample_names, file_paths)
}
\arguments{
\item{sample_names}{Vector of sample names to use}

\item{file_paths}{Vector of absolute path for each sample}
}
\value{
List of dataframes containing counts, taxonomy and all information for each sample
}
\description{
Get dragen metagenomics data for all samples
}
