% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_freq}
\alias{get_freq}
\title{Get the frequency of elements in a vector}
\usage{
get_freq(vect_)
}
\arguments{
\item{vect_}{Vector of elements}
}
\description{
Get the frequency of elements in a vector
}
