% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_functions.R
\name{get_gene_annotation}
\alias{get_gene_annotation}
\title{Get gene annotation from BioMart}
\usage{
get_gene_annotation(
  Genelist,
  species = "hsapiens_gene_ensembl",
  get_uniprot = FALSE
)
}
\arguments{
\item{GeneList}{Gene IDs entries. ENSMBL genes only.}

\item{datasets}{By default: hsapiens_gene_ensembl}
}
\description{
Gets annotation for the set of genes desired
}
