% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gene_coordinates}
\alias{get_gene_coordinates}
\title{Get gene coordinates from BioMart}
\usage{
get_gene_coordinates(Genelist, species = "hsapiens_gene_ensembl")
}
\arguments{
\item{GeneList}{Gene IDs entries. ENSMBL genes only.}

\item{datasets}{By default: hsapiens_gene_ensembl}
}
\description{
Gets coordinates for the set of genes desired
}
