% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_GSEA_datasets}
\alias{get_GSEA_datasets}
\title{Get GSEA Datasets}
\usage{
get_GSEA_datasets(species_given = "Homo sapiens")
}
\arguments{
\item{species_given}{Species to retrieve data. Homo sapiens by default}
}
\description{
Gets GSEA datasets and prepares data for FGSEA analyasis
}
