% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAcyto_functions.R
\name{get_info_GRanges}
\alias{get_info_GRanges}
\title{Get info for ASCAT}
\usage{
get_info_GRanges(row_id, df, tmp.list, ClinVar = FALSE)
}
\arguments{
\item{row_id}{Rowname id to subset}

\item{df}{ASCAT dataframe imported from file}

\item{tmp.list}{Temporary list, named tmp, that contains information per probe}

\item{ClinVar}{Boolean either if tmp.list is ClinVar or not}
}
\description{
Get info for ASCAT
}
