% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_functions.R
\name{get_length_cat}
\alias{get_length_cat}
\title{Get categories of gene length regions}
\usage{
get_length_cat(df_given)
}
\arguments{
\item{df_given}{Dataframe to use containing a length column}
}
\description{
Get categories of gene length regions
}
