% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_functions.R
\name{get_length_list_GRs}
\alias{get_length_list_GRs}
\title{Get length of regions for a list of GRs}
\usage{
get_length_list_GRs(list_GR)
}
\arguments{
\item{list_GR}{List of GRs to compute length}
}
\description{
Get length of regions for a list of GRs
}
