% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_length_vect}
\alias{get_length_vect}
\title{Get the length of unique elements in a vector}
\usage{
get_length_vect(vect_)
}
\arguments{
\item{vect_}{Vector of elements}
}
\description{
Get the length of unique elements in a vector
}
