% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_functions.R
\name{get_limma_EPIC_results}
\alias{get_limma_EPIC_results}
\title{Get limma results for EPIC probes}
\usage{
get_limma_EPIC_results(
  normData.given,
  sample_sheet.given,
  design.given,
  comp.name.given,
  folder_given,
  contrasts.matrix.here,
  illumina_annot,
  b_values_given,
  FCcutoff.given = log2(1.2),
  diff.mean.given = 0.2,
  paired_variable = NULL
)
}
\arguments{
\item{normData.given}{Dataframe with probes as rownames}

\item{sample_sheet.given}{Dataframe with information}

\item{design.given}{Statistical design}

\item{comp.name.given}{Name for the comparison variable}

\item{folder_given}{Folder to store results}

\item{contrasts.matrix.here}{Contrasts matrix to test}

\item{illumina_annot}{Illumina annotation dataframe}

\item{b_values_given}{Dataframe containing beta_values for each probe and sample (columns)}

\item{FCcutoff.given}{Log FC difference (absolute values) for filtering}

\item{mean.diff}{Mean difference (absolute values) for filtering}
}
\description{
Get limma results for EPIC probes
}
