% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_functions.R
\name{get_limma_results}
\alias{get_limma_results}
\title{Create limma DE analysis for all coefficients}
\usage{
get_limma_results(
  normData,
  df_treatment_Ind,
  design.given,
  contrasts.matrix.given,
  coef.given.list,
  OUTPUT_Data_sample,
  comp_name,
  pCutoff.given = 0.05,
  FCcutoff.given = log2(1.2),
  forceResults = FALSE
)
}
\arguments{
\item{normData}{dataframe to use as normalized values}

\item{df_treatment_Ind}{Sample sheet with information for plotting}

\item{design.given}{Matrix with design}

\item{contrasts.matrix.given}{makeContrast object to test}

\item{coef.given.list}{List of coefficients names to test.}

\item{OUTPUT_Data_sample}{Output directory to save results}

\item{comp_name}{Column name in sample sheet given to test.}

\item{pCutoff.given}{Adjusted pvlaue cutoff. Default=0.05,}

\item{FCcutoff.given}{Log Fold change cutoff. Default=log2(1.2),}

\item{forceResults}{Set flag to force results if previously generated. Default=FALSE}
}
\description{
This functions creates limma DE analysis
}
