% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen2phyloseq.R
\name{get_OTUs_dragen}
\alias{get_OTUs_dragen}
\title{Produce OTU matrix from dragen imported data}
\usage{
get_OTUs_dragen(dragen_data_samples)
}
\arguments{
\item{dragen_data_samples}{get_dragenData$data_samples result generated. A list of dataframes for eac sample}
}
\value{
Data frame with OTUs for each sample
}
\description{
Produce OTU matrix from dragen imported data
}
