% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_functions.R
\name{get_perc_overlap}
\alias{get_perc_overlap}
\title{Get percentage of overlaps for two GRs}
\usage{
get_perc_overlap(A, B, verbose = FALSE)
}
\arguments{
\item{A}{GenomicRanges A}

\item{B}{GenomicRanges B}

\item{verbose}{TRUE/FALSE for verbosity}
}
\description{
Only percentage of BinA will be reported
}
