% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAcyto_functions.R
\name{get_pfb}
\alias{get_pfb}
\title{Get PFB information}
\usage{
get_pfb(variant_data, AF2get = "Global.Allele.Frequency.from.1000G")
}
\arguments{
\item{variant_data}{Variant annotation file, provided by illumina}

\item{AF2get}{Allele frequency to obtain. As default: Global.Allele Freq. from 1k genomes}
}
\description{
Get Population frequency allele-B information for a given population
}
