% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_possible_comparisons}
\alias{get_possible_comparisons}
\title{Get all possible combinations for the levels provided}
\usage{
get_possible_comparisons(vec.given)
}
\arguments{
\item{vec.given}{Vector provided}
}
\description{
Get all possible combinations for the levels provided
}
