% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_functions.R
\name{get_prevalence}
\alias{get_prevalence}
\title{Get prevalence}
\usage{
get_prevalence(physeq.obj, results.fold, tag_name = "data", prevThres = 0.05)
}
\arguments{
\item{physeq.obj}{Phyloseq object to use}

\item{results.fold}{Results folder to store data in csv and xlsx format}

\item{tag_name}{Tag to add to the name of the files generated}

\item{prevThres}{Proportion of samples as threshold}
}
\description{
Get prevalence
}
