% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_functions.R
\name{get_probes_limma}
\alias{get_probes_limma}
\title{Create limma analysis for each probe}
\usage{
get_probes_limma(
  values_df,
  design_given,
  contrasts.matrix_given,
  pval_sign = 0.05,
  thr = 0.26
)
}
\arguments{
\item{values_df}{Values to test}

\item{design_given}{Statistical design}

\item{contrasts.matrix_given}{Contrasts to test}

\item{pval_sign}{Cufoff for pvalue adjusted}

\item{thr}{Threshold for LFC}
}
\description{
This functions creates a simplified version of limma analysis for each probe
}
