% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{get_Results_DDS}
\alias{get_Results_DDS}
\title{Generate results given a DDS object}
\usage{
get_Results_DDS(
  dds_object,
  OUTPUT_Data_dir_given,
  dfAnnotation,
  comp_ID,
  int_threads = 2,
  coef_n = NA
)
}
\arguments{
\item{dds_object}{DESeq2 object (DESeqDataSet)}

\item{OUTPUT_Data_dir_given}{Absolute path to store results}

\item{dfAnnotation}{Dataframe with useful metadata to include}

\item{comp_ID}{Tag name to include for each comparison}

\item{int_threads}{Number of threads to use in the analysis}

\item{coef_n}{Number of the coefficient of results to test \link{if desired}}
}
\description{
When running DESeq2 you usually add multiple terms to the matrix design. Test the effect of them
}
