% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpSetR_functions.R
\name{get_single_UpSet_plot}
\alias{get_single_UpSet_plot}
\title{Get UpSetR unique intersect}
\usage{
get_single_UpSet_plot(data_set, set2test)
}
\arguments{
\item{data_set}{Matrix containing information for each set and item. Created from: create_upset_data() and/or returned by create_upset()}

\item{set2test}{Set name to retrieve unique single items. It can be one or two items. e.g. male-female_down, c("example_down", "example-cond2_down")}
}
\value{
List of items unique for the set of interest
}
\description{
Get information of the unique items in a given intersect
}
