% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_functions.R
\name{get_sum_bp}
\alias{get_sum_bp}
\title{Get sum of bp between min and max length}
\usage{
get_sum_bp(i, min_length = 100, max_length = 1000)
}
\arguments{
\item{df_given}{Dataframe to use containing a length column}
}
\description{
Get sum of bp between min and max length
}
