% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_topGO_data}
\alias{get_topGO_data}
\title{Create topGO data and enrichment analysis}
\usage{
get_topGO_data(
  list_genes,
  list_all_genes,
  gene2GO,
  ont_given = "B",
  nodeSize_given = 5,
  cutoff_given = 0.05,
  statistic_given = "fisher"
)
}
\arguments{
\item{list_genes}{List of DE gene (EntrezID).}

\item{list_all_genes}{List of all available genes (EntrezID).}

\item{ont_given}{Ontology to analyse: Biological process (B); Mollecular function (M) or Celullar component (C)}

\item{nodeSize_given}{Size of the node used to prune the GO hierarchy from the terms which have less than X annotated genes}

\item{cutoff_given}{Pvalue cutoff given}

\item{gene2go}{ViSEAGO annotation object for the package annotation of interest.}
}
\description{
Creates Gene Ontology data using topGO package and enrichment analysis
}
