% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPICv2_functions.R
\name{getDataF_sesameQC_calcStats}
\alias{getDataF_sesameQC_calcStats}
\title{Dataframe generator from calcStats sesameQC}
\usage{
getDataF_sesameQC_calcStats(qcs.object)
}
\arguments{
\item{qcs.object}{sesameQC object generated}
}
\value{
dataframe
}
\description{
Given a sesameQC_calcStats qcs objects, create a dataframe and return it for all samples available
}
