% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{ggboxplot_scatter}
\alias{ggboxplot_scatter}
\title{Shorcut to using ggscatterBoxplots}
\usage{
ggboxplot_scatter(data_all_given, colName, y.coord)
}
\arguments{
\item{data_all_given}{Dataframe containing all information for all samples.}

\item{y.coord}{Dimension reduction coordinate (e.g. PC1, PC2), included as a column in data_all_given}

\item{colname}{Variable name to test. It should be included within data_all_given}
}
\description{
This code creates a ggscatter boxplot with t-test and Anova statistic by stat_compare
}
