% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{ggscatter_plotRegression}
\alias{ggscatter_plotRegression}
\title{Shorcut to using ggscatter regression plots}
\usage{
ggscatter_plotRegression(
  data_all_given,
  x.given,
  y.given,
  title_string = "",
  x_label = ""
)
}
\arguments{
\item{data_all_given}{Dataframe containing all information for all samples.}

\item{x.given}{Variable to plot in X axis. It should be numeric and included within data_all_given}

\item{y.given}{Variable to plot in Y axis. It should be numeric and included within data_all_given}

\item{title_string}{Title to add to the plot}

\item{x_label}{Title to add to the axis}
}
\description{
This code creates a ggscatter regression plot given two numeric variables.
}
