% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_functions.R
\name{gm_mean}
\alias{gm_mean}
\title{Geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{Matrix}

\item{na.rm}{Remove NAs}
}
\description{
Calculate geometric means prior to estimate size factors
}
