% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_functions.R
\name{gprofiler_analysis}
\alias{gprofiler_analysis}
\title{Create gprofiler gene ontology enrichment for a given dataset}
\usage{
gprofiler_analysis(
  genes_of_interest,
  dir2save,
  name2use,
  significant_ = TRUE,
  organism2use = "hsapiens"
)
}
\arguments{
\item{genes_of_interest}{List of genes (ENSEMBL ids)}

\item{name2use}{Name tag to add}

\item{significant_}{Flag to show only significant (default) or any result ranked by pvalue}

\item{organism2use}{Organism species name. Default: hsapiens}

\item{dirsave}{Folder to store results}
}
\description{
Create gprofiler gene ontology enrichment for a given dataset
}
