% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAcyto_functions.R
\name{GR_list_samples}
\alias{GR_list_samples}
\title{Create GR list subseting dataframe}
\usage{
GR_list_samples(data2use)
}
\arguments{
\item{data2use}{}
}
\description{
Create GR list subseting dataframe
}
