% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_functions.R
\name{GRanges_subsetter}
\alias{GRanges_subsetter}
\title{Subset lisf ot GRanges objects}
\usage{
GRanges_subsetter(list_of_GRanges, sub_Granges = NULL, string_given = NULL)
}
\arguments{
\item{list_of_GRanges}{}

\item{sub_Granges}{GRanges to use to subset}

\item{string_given}{String to use to subset}
}
\description{
Subset lisf ot GRanges objects
}
