% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_functions.R
\name{limma_DE_function}
\alias{limma_DE_function}
\title{Create limma DE analysis}
\usage{
limma_DE_function(
  efit_3,
  normData,
  df_treatment_Ind,
  design.given,
  contrasts.matrix.given,
  coef.given,
  OUTPUT_Data_sample,
  comp_name,
  numerator,
  denominator,
  pCutoff.given = 0.05,
  FCcutoff.given = log2(1.2),
  forceResults = FALSE,
  EPIC = FALSE
)
}
\arguments{
\item{efit_3}{Ebayes fit results}

\item{normData}{dataframe to use as normalized values}

\item{df_treatment_Ind}{Sample sheet with information for plotting}

\item{design.given}{Matrix with design}

\item{contrasts.matrix.given}{makeContrast object to test}

\item{coef.given}{One coefficient to test.}

\item{OUTPUT_Data_sample}{Output directory to save results}

\item{comp_name}{Column name in sample sheet given to test.}

\item{numerator}{Comparison numerator.}

\item{denominator}{Comparison denominator.}

\item{pCutoff.given}{Adjusted pvlaue cutoff. Default=0.05,}

\item{FCcutoff.given}{Log Fold change cutoff. Default=log2(1.2),}

\item{forceResults}{Set flag to force results if previously generated. Default=FALSE}

\item{EPIC}{Boolean to indicate EPIC probes, avoid generating huge xlsx file. Default=FALSE}
}
\description{
This functions creates limma DE analysis
}
