% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_annotation_EPICv2.R
\name{load_annotation_EPICv2}
\alias{load_annotation_EPICv2}
\title{Load different objects into a GenomicRanges list}
\usage{
load_annotation_EPICv2(
  EPIC_annotv2 =
    "/imppc/labs/lslab/share/data/references/EPIC/EPIC_v2/EPICv2.annot.RData",
  re_annotated =
    "/imppc/labs/lslab/share/data/references/EPIC/EPIC_v2/EPICv2_reannotated_manifest_v2.0.csv",
  elements_UCSC =
    "/imppc/labs/lslab/share/data/references/EPIC/EPIC_v2/elements_UCSC_annot.RData",
  gtf_file =
    "/imppc/labs/lslab/share/data/references/Human_reference_latest/Homo_sapiens.GRCh38.108.chr._ncRNA-RNAcentral.gtf"
)
}
\arguments{
\item{EPIC_annotv2}{}

\item{re_annotated}{}

\item{elements_UCSC}{}

\item{gtf_file}{}
}
\description{
Load different objects into a GenomicRanges list
}
