% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBeads_functions.R
\name{load_RnBeads_object}
\alias{load_RnBeads_object}
\title{Loads RnBead object}
\usage{
load_RnBeads_object(dir4loading)
}
\arguments{
\item{dir4loading}{Absolute path to folder to load RnBeads item}
}
\description{
This functions loads a previously saved RnBeads object in a file and dir.
}
