% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loader}
\alias{loader}
\title{Loads R data into variable}
\usage{
loader(file2load)
}
\arguments{
\item{file2load}{Absolute path to save RData file}
}
\description{
This functions loads a given RData object in a temporal environment and returns it
}
