% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBeads_functions.R
\name{locus_plotter}
\alias{locus_plotter}
\title{RnBeads Locus plotter caller}
\usage{
locus_plotter(
  RnBeads_object,
  grouping_class,
  gene_name,
  comparison,
  outdir,
  chr,
  start,
  end
)
}
\arguments{
\item{RnBeads_object}{RnBeads object}

\item{grouping_class}{rnb.sample.groups output grouping samples}

\item{gene_name}{Name of the gene/region to create pdf}

\item{comparison}{Name to add to create pdf}

\item{outdir}{Folder to store pdf}

\item{chr}{Chromosome ID}

\item{start}{Start position}

\item{end}{End position}
}
\description{
This functions calls rnb.plot.locus.profile for a gene or region of interest
}
