% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{make_all_contrasts}
\alias{make_all_contrasts}
\title{Make all pairwise contrast for limma}
\usage{
make_all_contrasts(group, delim = "_vs_", decreasing = TRUE)
}
\arguments{
\item{group}{Vector of names to use}

\item{delim}{Character to use for delimiting: "\emph{vs}" as default}
}
\description{
Original: https://bioinformatics.stackexchange.com/questions/18570/generating-contrast-matrix-for-limma-in-loop
}
