% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{maplot}
\alias{maplot}
\title{Create MA plot}
\usage{
maplot(res, thresh = 0.05, textcex = 1, ...)
}
\arguments{
\item{res}{dataframe}

\item{thresh}{pvalue adjusted threshold}

\item{textcex}{determines text size}
}
\description{
This functions creates MA plots given a dataframe containing:
baseMean, log2FoldChange and padj header names.
}
