% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBeads_functions.R
\name{my_volcanoplot_region}
\alias{my_volcanoplot_region}
\title{VolcanoPlot region}
\usage{
my_volcanoplot_region(
  res,
  lfcthresh = 2,
  sigthresh = 0.05,
  main = "Volcano Plot",
  legendpos = "topleft",
  textcx = 1,
  point_cex = 0.3,
  ...
)
}
\description{
This functions creates volcano plot using a region of interest
}
