% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAcyto_functions.R
\name{pheatmap_log}
\alias{pheatmap_log}
\title{Create pheatmap using log values}
\usage{
pheatmap_log(matrixinp_given, meta_data_given = NULL)
}
\arguments{
\item{matrixinp_given}{}

\item{meta_data_given}{}
}
\description{
Create pheatmap using log values
}
