% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot.grid.ggscatterBoxplot}
\alias{plot.grid.ggscatterBoxplot}
\title{Grid plots using ggscatterBoxplots}
\usage{
\method{plot}{grid.ggscatterBoxplot}(list2test, data_all_given, coord.char)
}
\arguments{
\item{list2test}{List of columns names to test. They should be valid names and should be included within data_all_given}

\item{data_all_given}{Dataframe containing all information for all samples.}

\item{coord.char}{Dimension reduction coordinate (e.g. PC1, PC2), included as a column in data_all_given}
}
\description{
This code create multiple plots using ggboxplot_scatter for a list of variables, given
a coordinate (PC1, PC2, UMAP1, UMAP2...). All variables and coordinate should be included in
the data_all_given provided
}
