% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_batch_effect}
\alias{plot_batch_effect}
\title{Plot batch effect}
\usage{
plot_batch_effect(var1, var2, dds_object, dirName, batch_var)
}
\arguments{
\item{var1}{DESeq2 object}

\item{var2}{DESeq2 object}

\item{dds_object}{DESeq2 object}

\item{dirName}{Folder path to store results}

\item{batch_var}{Putative batch variable}
}
\description{
This functions plots original PCA and batch corrected given two variables and a putative batch variable
}
