% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meffil_functions.R
\name{plot_control_means}
\alias{plot_control_means}
\title{Function to create single plots by control probe category}
\usage{
plot_control_means(
  data_plot,
  vect_categories,
  metadata,
  col_2_color,
  col_of_interest
)
}
\arguments{
\item{data_plot}{Dataframe from the original ggplot in meffil qc.summary (qc.summary_given$controlmeans.summary)}

\item{vect_categories}{Vector containing list of categories to plot}

\item{metadata}{Dataframe with metadata associated to each sample}

\item{col_2_color}{Column to color ggplot points}

\item{col_of_interest}{Columns from metadata of interest}
}
\description{
Function to create single plots by control probe category
}
