% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{plot_DESeq2_pvalues}
\alias{plot_DESeq2_pvalues}
\title{Plot DESeq2 p-values}
\usage{
plot_DESeq2_pvalues(pdf_name, res, res_filtered)
}
\arguments{
\item{pdf_name}{File path for the pdf to save images}

\item{res}{Original results (dds object)}

\item{res_filtered}{Filtered results (dds object)}
}
\description{
This functions plots pvalues original and after filtering
}
