% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_functions.R
\name{plot_gene_values}
\alias{plot_gene_values}
\title{Plot values per genes}
\usage{
plot_gene_values(gene, tableCounts, targetsFile, condition, out_folder)
}
\arguments{
\item{gene}{Gene ID}

\item{tableCounts}{Expression counts. Samples as columns}

\item{targetsFile}{Phenothypic information. Sampes as rows.}

\item{condition}{List of phenotypic condition to retrieve in targestFile}

\item{out_folder}{Out folder to create plot}
}
\description{
Plot DESeq2 normalization values using ggplot2
}
