% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_GSEA}
\alias{plot_GSEA}
\title{Plog GSEA}
\usage{
plot_GSEA(fgRes, title_given)
}
\arguments{
\item{fgRes}{Dataframe containing information}

\item{title_given}{Title to include in the ggplot generated}
}
\description{
Plots GSEA results using ggplot and enrichment score provided
}
