% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPICv2_functions.R
\name{plot_identity.bar}
\alias{plot_identity.bar}
\title{Plot identity bars}
\usage{
plot_identity.bar(df.given)
}
\arguments{
\item{df.given}{Dataframe from getDataF_sesameQC_calcStats}
}
\description{
Plot identity bars
}
