% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_representation}
\alias{plot_representation}
\title{Create PCA/UMAP plot}
\usage{
plot_representation(
  row_names,
  x_given,
  y_given,
  meta_data_col,
  title,
  label_plot = FALSE,
  size = FALSE,
  size_var = NULL,
  max.overlaps_int = 50,
  encircle = FALSE
)
}
\arguments{
\item{row_names}{Identifiers for each sample}

\item{x_given}{Variable to plot in X axis. It should be numeric and included within data_all_given}

\item{y_given}{Variable to plot in Y axis. It should be numeric and included within data_all_given}

\item{meta_data_col}{Vector for colouring}

\item{title}{Title to add to the plot}

\item{label_plot}{Add labels to the points}

\item{size}{TRUE/FALSE to include a numeric variable with different size points}

\item{size_var}{numeric variable with different size points}

\item{max.overlaps_int}{Number of overlaps for ggrepel}

\item{encircle}{Wether to include circles aroung groups of row_names provided}
}
\description{
This functions creates a PCA or UMAP plot given a vector of names, x and y coordinates. You can additionally
colour according to other vector, add title, label the points, set a different variabel to
include and additional parameter (size), and create circles aroung groups of row_names provided.
}
