% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAcyto_functions.R
\name{plotCN.here}
\alias{plotCN.here}
\title{Extended genoCN plot for BAF and LRR}
\usage{
plotCN.here(
  pos,
  LRR,
  BAF,
  chr2plot = NULL,
  sampleIDs = NULL,
  plot_names = "genoCN",
  fileNames = NULL,
  types = "genoCN",
  CNA = TRUE,
  main = "",
  LRR.ylim = NULL,
  cex = 0.5,
  plot.lowess = TRUE
)
}
\arguments{
\item{pos}{}

\item{LRR}{}

\item{BAF}{}

\item{chr2plot}{}

\item{sampleIDs}{}

\item{plot_names}{}

\item{fileNames}{}

\item{types}{}

\item{CNA}{}

\item{main}{}

\item{LRR.ylim}{}

\item{cex}{}

\item{plot.lowess}{}
}
\description{
Extended genoCN plot for BAF and LRR
}
