% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAcyto_functions.R
\name{plotCN_zoom}
\alias{plotCN_zoom}
\title{Extended genoCN plot for BAF and LRR using a Zoom region}
\usage{
plotCN_zoom(
  segment_file,
  snpInfo.subset.df,
  sample2process_data,
  start_bp = 0,
  stop_bp = NULL,
  several_tracks = NULL,
  types.given = c("genoCN", "genoCN"),
  main_title = "example plot",
  plot_names.here = NULL,
  verbose = FALSE
)
}
\arguments{
\item{segment_file}{}

\item{snpInfo.subset.df}{}

\item{sample2process_data}{}

\item{start_bp}{}

\item{stop_bp}{}

\item{several_tracks}{}

\item{types.given}{}

\item{main_title}{}

\item{plot_names.here}{}

\item{verbose}{}
}
\description{
Extended genoCN plot for BAF and LRR using a Zoom region
}
