% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meffil_functions.R
\name{print_plot_control_means}
\alias{print_plot_control_means}
\title{Create control probe plots}
\usage{
print_plot_control_means(
  data_plot,
  vect_categories,
  col_2_color,
  tag_to_use,
  plots_dir,
  col_of_interest,
  metadata,
  max_cols = 2,
  max_rows = 2
)
}
\arguments{
\item{data_plot}{Dataframe from the original ggplot in meffil qc.summary (qc.summary_given$controlmeans.summary)}

\item{vect_categories}{Vector containing list of categories to plot}

\item{col_2_color}{Column to color ggplot points}

\item{tag_to_use}{String to include at the beggining of the PDF named generated}

\item{plots_dir}{Absolute path to folder to store results}

\item{col_of_interest}{Columns from metadata of interest}

\item{metadata}{Dataframe with metadata associated to each sample}

\item{max_cols}{Number of columns of the grid of plots}

\item{max_rows}{Number of rows of the grid of plots}
}
\description{
Create control probe plots
}
