% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meffil_functions.R
\name{produce_norm_plots}
\alias{produce_norm_plots}
\title{Improve norm plots generated by meffil}
\usage{
produce_norm_plots(
  set_of_plots,
  name_set,
  tag_to_use_here,
  meffil_plots_dir_given
)
}
\arguments{
\item{set_of_plots}{List of plots, either control.batch or probe.batch containing fplots, cplots and pc.plots}

\item{name_set}{Name of the list provided: control.batch or probe.batch}

\item{tag_to_use_here}{Tag to add to the file}

\item{meffil_plots_dir_given}{Folder to store results}
}
\description{
Improve norm plots generated by meffil
}
