% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPICv2_functions.R
\name{pval_maskplot}
\alias{pval_maskplot}
\title{Pvalue mask filtering}
\usage{
pval_maskplot(SigDF_given, sample_name, sample_dir, tag_name = "raw_")
}
\arguments{
\item{sample_name}{Sample ID}

\item{sample_dir}{Folder to store results}

\item{tag_name}{Tag to add to the files generated}

\item{sample_frame}{SigDF generated}
}
\description{
This function creates an histogram overlaid with kernel density curve
It uses the sesame::pOOBAH function to retrieved the pvalue detection
}
