% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rank_list_by}
\alias{rank_list_by}
\title{Rank table by given variable}
\usage{
rank_list_by(table_data, option_given = "logFC")
}
\arguments{
\item{table_data}{Dataframe containing information}

\item{option_given}{Column name to retrieve data.}
}
\description{
Ranks a given table by the given column name. It requires a GENE_SYMBOL column to include gene IDs.
}
