% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAcyto_functions.R
\name{read_ASCAT}
\alias{read_ASCAT}
\title{Read ASCAT output file}
\usage{
read_ASCAT(
  ascat_file,
  list_Granges,
  list_Granges.ClinVar,
  verbose = FALSE,
  test_n = NULL
)
}
\arguments{
\item{genoCN_file}{}
}
\description{
Read ASCAT output file
}
