% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen2phyloseq.R
\name{read_dragen_sample_summary}
\alias{read_dragen_sample_summary}
\title{Read DRAGEN summary information}
\usage{
read_dragen_sample_summary(
  file_sample,
  nameGiven,
  tax_levels = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")
)
}
\arguments{
\item{file_sample}{Absolute path to file}

\item{nameGiven}{Name to use for sample identification}

\item{tax_levels}{Taxonomic levels to import. Default: c("Kingdom","Phylum","Class","Order","Family","Genus", "Species")}
}
\value{
Dataframe
}
\description{
For each sample, DRAGEN metagenomics creates a summary file (csv)
that contains counts per species that can be added at any taxonomic level
of interest. Here we read and prepare for later merging or additional analysis
}
