% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_functions.R
\name{relevel_function}
\alias{relevel_function}
\title{Relevel and rung DESEQ2 analysis}
\usage{
relevel_function(
  dds_object,
  category,
  reference,
  given_dir,
  dfAnnotation,
  list_of_cols,
  data_type = "mRNA",
  gene.annot.df.given = NULL,
  int_threads = 2,
  sign_value.given = 0.05,
  LFC.given = log2(1.2),
  min_cutoff_to_plot = 3,
  max_cutoff_to_plot = 50,
  comp_ID.given = "comp1",
  forceResults = FALSE,
  localFit = FALSE,
  shrinkage.given = "apeglm"
)
}
\arguments{
\item{dds_object}{DESeq2 object}

\item{category}{Name of the variable to use within dds_object metadata}

\item{reference}{Name of the reference class to set.}

\item{given_dir}{Output dir to use}

\item{dfAnnotation}{Dataframe containing additional information for each sample}

\item{data_type}{Either mRNA, miRNA or rRNA_16S. Default: mRNA}

\item{int_threads}{Number of CPUs to use (Default: 2)}

\item{sign_value.given}{Adjusted pvlaue cutoff. Default=0.05,}

\item{LFC.given}{Log Fold change cutoff. Default=log2(1.2),}

\item{forceResults}{Boolean to force re-run analysis if already generated in the folder provided}

\item{localFit}{Use a fitType=local for mean dispersion fit in DESeq2}

\item{shrinkage.given}{LFC shrinkage estimator provided. Available: apeglm, ashr or normal}
}
\description{
When running DESeq2 you sometimes require to relevel some comparisons
}
