% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_functions.R
\name{save_bed_info}
\alias{save_bed_info}
\title{Save BUMP information in BED file format}
\usage{
save_bed_info(bump_table, file_out)
}
\arguments{
\item{bump_table}{Bumphunter dataframe}

\item{file_out}{File to store results}
}
\description{
This functions saves a bump_table provided in BED format for later analysis
}
