% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_jpeg}
\alias{save_jpeg}
\title{Save plots in JPEG format}
\usage{
save_jpeg(folder_path, name_file, plot_given)
}
\arguments{
\item{folder_path}{}

\item{name_file}{}

\item{plot_given}{}
}
\description{
Save plots in JPEG format
}
