% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_functions.R
\name{save_multi_pdf}
\alias{save_multi_pdf}
\title{Save multiple plots in PDF}
\usage{
save_multi_pdf(folder_path, name_file, list_plots)
}
\arguments{
\item{folder_path}{Folder to store results}

\item{name_file}{Name of the file}

\item{list_plots}{List of plots to save in pdf}
}
\description{
Save multiple plots in PDF
}
