% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meffil_functions.R
\name{save_norm_plots}
\alias{save_norm_plots}
\title{Save and improve norm plots generated by meffil}
\usage{
save_norm_plots(
  meffil_plots_dir_given,
  tag_to_use_given,
  norm.summary_given,
  meffil_object_dir_given
)
}
\arguments{
\item{meffil_plots_dir_given}{}

\item{tag_to_use_given}{}

\item{norm.summary_given}{}

\item{meffil_object_dir_given}{}
}
\description{
Save and improve norm plots generated by meffil
}
