% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_pdf}
\alias{save_pdf}
\title{Save plot in pdf}
\usage{
save_pdf(folder_path, name_file, plot_given)
}
\arguments{
\item{folder_path}{Absolute path to save pdf}

\item{name_file}{PDF file name, do not include extension}

\item{plot_given}{Plot object}
}
\description{
Save a plot in a pdf in the folder provided.
}
