% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_plots_multiformat}
\alias{save_plots_multiformat}
\title{Save plots in multiple formats}
\usage{
save_plots_multiformat(folder_path, name_file, plot_given)
}
\arguments{
\item{folder_path}{}

\item{name_file}{}

\item{plot_given}{}
}
\description{
Save plots in multiple formats
}
