% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_png}
\alias{save_png}
\title{Save plots in PNG format}
\usage{
save_png(folder_path, name_file, plot_given)
}
\arguments{
\item{folder_path}{}

\item{name_file}{}

\item{plot_given}{}
}
\description{
Save plots in PNG format
}
