% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meffil_functions.R
\name{save_qc_plots}
\alias{save_qc_plots}
\title{Save QC summary plots as PDF}
\usage{
save_qc_plots(
  meffil_plots_dir_given,
  tag_to_use_given,
  qc.summary_given,
  metadata_given,
  col_of_interest
)
}
\arguments{
\item{meffil_plots_dir_given}{Folder path to store plots}

\item{tag_to_use_given}{String to add to each file}

\item{qc.summary_given}{QC summary object generated}
}
\description{
Save QC summary plots as PDF
}
