% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBeads_functions.R
\name{save_RnBeads_object}
\alias{save_RnBeads_object}
\title{Saves RnBead object}
\usage{
save_RnBeads_object(RnBeads_object, file2write, dir2save)
}
\arguments{
\item{RnBeads_object}{RnBeads object to save}

\item{file2write}{File name to use}

\item{dir2save}{Absolute path to folder to save item}
}
\description{
This functions saves a given RnBeads object in a file and dir provided
}
