% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_functions.R
\name{save_woorkbook_openxlsx}
\alias{save_woorkbook_openxlsx}
\title{Save multiple dataframes in XLSX}
\usage{
save_woorkbook_openxlsx(file_path, list_df)
}
\arguments{
\item{folder_path}{File to store results (.xlsx)}

\item{list_plots}{List of dataframes to save. Names as tab names.}
}
\description{
Save multiple dataframes in XLSX
}
