% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sessionInfo_write}
\alias{sessionInfo_write}
\title{Saves R sessionInfo data into file}
\usage{
sessionInfo_write(dir.given)
}
\arguments{
\item{dir.given}{Absolute path to save file}
}
\description{
This functions saves the configuration of the session into a txt file
}
