% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_functions.R
\name{stats_GR}
\alias{stats_GR}
\title{Get count of region lengths for each GR}
\usage{
stats_GR(
  given_GR,
  length2use = c(1000, 10000, 1e+05, 1e+06, 1e+07, 1e+08),
  names2use = c("1kbp", "10kbp", "100kbp", "1Mbp", "10Mbp", "100Mbp")
)
}
\arguments{
\item{A}{GenomicRanges A}

\item{B}{GenomicRanges B}

\item{verbose}{TRUE/FALSE for verbosity}
}
\description{
Get count of region lengths for each GR
}
