% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meffil_functions.R
\name{vector_splitter}
\alias{vector_splitter}
\title{Split vector into parts}
\usage{
vector_splitter(vector_of_names, max_cols = 2, max_rows = 2)
}
\arguments{
\item{vector_of_names}{Vector of items}

\item{max_cols}{Number of columns}

\item{max_rows}{Number of rows}
}
\description{
Given a vector, splitted in many parts with equal number of items for later creating grid of plots
}
