% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{volcanoplot}
\alias{volcanoplot}
\title{Create Volcano plot}
\usage{
volcanoplot(
  res,
  lfcthresh = 2,
  sigthresh = 0.05,
  main = "Volcano Plot",
  legendpos = "topright",
  textcx = 1,
  ...
)
}
\arguments{
\item{res}{dataframe}

\item{lfcthresh}{Log fold change threshold}

\item{sigthresh}{pvalue adjusted threshold}

\item{main}{Main plot legend}

\item{legendpos}{Legend position}
}
\description{
This functions creates volcano plots given a dataframe containing:
baseMean, log2FoldChange and padj header names.
}
