% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meffil_functions.R
\name{write_norm_files}
\alias{write_norm_files}
\title{Save txt file for normalization objects}
\usage{
write_norm_files(norm.summary_given, meffil_data_dir_given, tag_to_use_here)
}
\arguments{
\item{norm.summary_given}{}

\item{meffil_data_dir_given}{}

\item{tag_to_use_here}{}
}
\description{
Save txt file for normalization objects
}
