% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meffil_functions.R
\name{write_qc_files}
\alias{write_qc_files}
\title{Function to save txt information from meffil QC analysis}
\usage{
write_qc_files(qc.summary_given, meffil_data_dir_given, tag_to_use_here)
}
\arguments{
\item{qc.summary_given}{QC summary object generated}

\item{meffil_data_dir_given}{Folder path to store files}

\item{tag_to_use_here}{String to add to each file}
}
\description{
Function to save as txt files all information generated during QC analysis from meffil
}
