% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{analysis_DESeq}
\alias{analysis_DESeq}
\title{DESEQ2 analysis pipeline}
\usage{
analysis_DESeq(
  OUTPUT_Data_dir_given,
  count_table,
  sample_sheet_given,
  dfAnnotation,
  int_threads,
  formula_given,
  coef_n = NA,
  early_return = FALSE,
  comp_ID = NULL
)
}
\arguments{
\item{OUTPUT_Data_dir_given}{Absolute path to store results}

\item{count_table}{Dataframe/matrix of counts}

\item{sample_sheet_given}{Samplesheet containing metadata information}

\item{dfAnnotation}{Dataframe with useful metadata to include}

\item{int_threads}{Number of threads to use}

\item{formula_given}{Design formula to use}

\item{coef_n}{Number of the coefficient of results to test \link{if desired}}

\item{early_return}{Whether to return exploratory results early or not}

\item{comp_ID}{Tag name to include for each comparison}
}
\description{
When running DESeq2 you usually add multiple terms to the matrix design. Test the effect of them
}
