% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBeads_functions.R
\name{beta_values_plots}
\alias{beta_values_plots}
\title{BetaDistribution plots}
\usage{
beta_values_plots(rnb.set_given, variablesOfInterest, pdf_name)
}
\description{
This functions creates beta distribution plots for a set of variables of interest
}
