% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_functions.R
\name{check_reduced_LRT}
\alias{check_reduced_LRT}
\title{Check the effect of reducing one term from a DESEQ2 design}
\usage{
check_reduced_LRT(
  dds_obj.given,
  formula_given,
  comp.folder.given,
  compID.given,
  dfAnnotation.given,
  list_of_cols,
  int_threads = 2,
  sign_value.given = 0.05,
  LFC.given = log2(1.2),
  min_cutoff_to_plot = 3,
  max_cutoff_to_plot = 50,
  gene.annot = NULL,
  forceResults = FALSE,
  shrinkage.given = "apeglm",
  data_type = "mRNA"
)
}
\arguments{
\item{dds_obj.given}{DESeq2 object (DESeqDataSet)}

\item{formula_given}{Formula to substract for formula in dds_obj.given}

\item{comp.folder.given}{Absolute path to store results}

\item{compID.given}{Tag name to include for each comparison}

\item{dfAnnotation.given}{Dataframe with useful metadata to include}

\item{forceResults}{Boolean to force re-run analysis if already generated in the folder provided}

\item{shrinkage.given}{LFC shrinkage estimator provided. Available: apeglm, ashr or normal}

\item{data_type}{Either mRNA, miRNA or rRNA_16S. Default: mRNA}
}
\description{
Check reducing effect using LRT method.
}
