% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_functions.R
\name{create_report}
\alias{create_report}
\title{Annotates GR Report for each region of bumps provided}
\usage{
create_report(bump_table, out_report, df_chr)
}
\arguments{
\item{bump_table}{Bumphunter dataframe}

\item{out_report}{Not used anymore, to be discarded}

\item{df_chr}{Dataframe generated as in get_chr_len() containing names and length for each Chromosome of interest}
}
\description{
This functions create a GR object with the regions of interest from bump_table provided
}
