% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpSetR_functions.R
\name{create_upset_plot}
\alias{create_upset_plot}
\title{Plots UpSetR}
\usage{
create_upset_plot(
  data_set,
  sets,
  y.label = "Items shared",
  x.label = "Items/group"
)
}
\arguments{
\item{data_set}{A dataframe containing as columns de sets, as rows the items and values are 0/1 for absence/presence. You can also provide create_upset_data(list_files)}

\item{sets}{Names of the set of interest to include in the comparison}

\item{y.label}{Name to include in the Y axis. Default: Items shared}

\item{x.label}{Name to include in the X axis. Default: Items/group}
}
\value{
Returns plot and dataset used
}
\description{
Creates plot with multiple interesection between sets using UpSetR package
}
