% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{DESeq2_HCGB_function}
\alias{DESeq2_HCGB_function}
\title{Plot DESeq2 p-values}
\usage{
DESeq2_HCGB_function(
  dds_object,
  coef_n,
  comp_name,
  comp_ID = "comp1",
  numerator = "example1",
  denominator = "example2",
  OUTPUT_Data_dir,
  df_treatment_Ind,
  threads = 2
)
}
\arguments{
\item{dds_object}{DESeq2 object}

\item{coef_n}{Coefficient number obtain using resultsName(dds)}

\item{numerator}{Name of the numerator comparison}

\item{denominator}{Name of the denominator comparison}

\item{OUTPUT_Data_dir}{Folder path to store results}

\item{df_treatment_Ind}{Dataframe containing additional information for each sample}

\item{threads}{Number of CPUs to use \link{Default: 2}.}

\item{name}{Name of the comparison}
}
\description{
This functions plots pvalues original and after filtering
}
