% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{exploratory_plots}
\alias{exploratory_plots}
\title{Exploratory plots for DESEQ2 analysis}
\usage{
exploratory_plots(dds_object.exp, OUTPUT_dir, dfAnnotation_df, list_of_cols)
}
\arguments{
\item{dds_object.exp}{DESeq2 object (DESeqDataSet)}

\item{OUTPUT_dir}{Absolute path to store results}

\item{dfAnnotation_df}{Dataframe with useful metadata to include}

\item{list_of_cols}{List of columns of interest to subset from metadata}
}
\description{
When running DESeq2 you usually add multiple terms to the matrix design. Test the effect of them
}
