% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_functions.R
\name{get_chr_len}
\alias{get_chr_len}
\title{Return dataframe with chromosome length for hg19 genome}
\usage{
get_chr_len(hg19 = TRUE)
}
\arguments{
\item{hg19}{Default. Only option available}
}
\description{
This functions returns chromosome length for hg19 genome
}
