% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_functions.R
\name{get_Results_DDS}
\alias{get_Results_DDS}
\title{Generate results given a DDS object}
\usage{
get_Results_DDS(
  dds_object,
  OUTPUT_Data_dir_given,
  dfAnnotation,
  list_of_cols,
  comp_ID,
  sign_value.given = 0.05,
  LFC.given = log2(1.2),
  int_threads = 2,
  coef_n = NA,
  forceResults = FALSE,
  gene.annot = NULL,
  shrinkage.given = "apeglm"
)
}
\arguments{
\item{dds_object}{DESeq2 object (DESeqDataSet)}

\item{OUTPUT_Data_dir_given}{Absolute path to store results}

\item{dfAnnotation}{Dataframe with useful metadata to include}

\item{comp_ID}{Tag name to include for each comparison}

\item{sign_value.given}{Adjusted pvlaue cutoff. Default=0.05,}

\item{LFC.given}{Log Fold change cutoff. Default=log2(1.2),}

\item{int_threads}{Number of threads to use in the analysis}

\item{coef_n}{Number of the coefficient of results to test \link{if desired}}

\item{forceResults}{Boolean to force re-run analysis if already generated in the folder provided}

\item{gene.annot}{Dataframe containing gene annotation (Default: NULL)}

\item{shrinkage.given}{LFC shrinkage estimator provided. Available: apeglm, ashr or normal}
}
\description{
When running DESeq2 you usually add multiple terms to the matrix design. Test the effect of them
}
