% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_functions.R
\name{limma_DE_function}
\alias{limma_DE_function}
\title{Create limma DE analysis}
\usage{
limma_DE_function(normData, design.given, contrasts.matrix.given, coef.given)
}
\arguments{
\item{normData}{dataframe to use as normalized values}

\item{design.given}{Matrix with design}

\item{contrasts.matrix.given}{makeContrast object to test}

\item{coef.given}{List of coefficients names to test.}
}
\description{
This functions creates limma DE analysis
}
