% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_functions.R
\name{litfover_func}
\alias{litfover_func}
\title{Convert hg19 <-> hg38 coordinates}
\usage{
litfover_func(
  folderWithLiftoverInfo,
  data2convert,
  hg19ToHg38 = FALSE,
  hg38ToHg19 = FALSE
)
}
\arguments{
\item{folderWithLiftoverInfo}{Folder containing liftover chain information retrieved from UCSC website}

\item{hg19ToHg38}{Boolean to transfrom hg19 to hg38}

\item{hg38ToHg19}{Boolean to transfrom hg38 to hg19}

\item{df2convert}{GenomicRanges or coordinate Dataframe provided to convert to GenomicRanges}
}
\description{
Using rtracklayer and GenomicRanges, this function converts hg19 <-> hg38 coordinates
}
