% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{relevel_function}
\alias{relevel_function}
\title{Relevel and rung DESEQ2 analysis}
\usage{
relevel_function(
  dds_object,
  category,
  reference,
  given_dir,
  dfAnnotation,
  int_threads = 2
)
}
\arguments{
\item{dds_object}{DESeq2 object}

\item{category}{Name of the variable to use within dds_object metadata}

\item{reference}{Name of the reference class to set.}

\item{given_dir}{Output dir to use}

\item{dfAnnotation}{Dataframe containing additional information for each sample}

\item{int_threads}{Number of CPUs to use \link{Default: 2}.}
}
\description{
When running DESeq2 you sometimes require to relevel some comparisons
}
